# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TS-API Documantation",
    "description": "In this documantation you will be able to consult the API end-points and also test all availblw routes.",
    "contact": {
      "email": "joey_qpb@hotmail.com"
    },
    "version": "1.0.0"
  },
  "serves": [
    {
      "url": "http://localhost:3333/TS-API",
      "description": "Local Server"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "get all users",
        "tags": ["/Users"],
        "description": "this route gives you access to all Users",
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/users"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Something went wrong."
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "get a user",
        "tags": ["/Users"],
        "description": "this route gives you access to a user data",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "user id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "404": {
            "description": "not found"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },
    "/create_user": {
      "post": {
        "summary": "Create a new user",
        "tags": ["/Users"],
        "description": "it will be responsible for creating a new User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              },
              "examples": {
                "New user": {
                  "value": {
                    "name": "joey",
                    "email": "joey@ex.com",
                    "telefone": 77999862226,
                    "enderço": {
                      "cidade": "VCA",
                      "cep": 450000,
                      "rua": "rua dos bobos",
                      "numero": 0
                    },
                    "cpf": "00000000000"
                  }
                },
                "Wrong New user": {
                  "value": {
                    "name": "joey",
                    "email": "joey@ex.com",
                    "telefone": 77999862226
                  }
                },
                "Wrong New user 2": {
                  "value": {
                    "name": "joey",
                    "email": "joey@ex.com",
                    "telefone": 77999862226,
                    "enderço": {
                      "cidade": "VCA",
                      "cep": 450000,
                      "rua": "rua dos bobos",
                      "numero": 0
                    },
                    "cpf": "00000000000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "car created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "400": {
            "description": "Specify a Body."
          },
          "500": {
            "description": "somenthing went wrong..."
          }
        }
      }
    },
    "/edit_user/{id}": {
      "patch": {
        "summary": "edit user",
        "tags": ["/Users"],
        "description": "this route allows to change the specific User registered in the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "Users id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/edit_User"
              },
              "examples": {
                "User edited": {
                  "value": {
                    "email": "newemail@ex.com",
                    "enderço": {
                      "cidade": "New City",
                      "cep": 2020202,
                      "rua": "New Street",
                      "numero": 1010
                    },
                    "telefone": 777
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "user edited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    },
    "/delete_user/{id}": {
      "delete": {
        "summary": "delete user",
        "tags": ["/Users"],
        "description": "delete the specific user registered in the database",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "user id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "user deleted",
            "content": {
              "application/json": {
                "schema": {
                  "acknowledged": true,
                  "deletedCount": 1
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "API error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "users": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "telefone": { "type": "number" },
          "enderço": {
            "type": "object",
            "properties": {
              "cidade": { "type": "string" },
              "cep": { "type": "number" },
              "rua": { "type": "string" },
              "numero": { "type": "number" }
            }
          },
          "cpf": { "type": "string" }
        }
      },
      "edit_User": {
        "email": { "type": "string" },
        "enderço": {
          "type": "object",
          "properties": {
            "cidade": { "type": "string" },
            "cep": { "type": "number" },
            "rua": { "type": "string" },
            "numero": { "type": "number" }
          }
        },
        "telefone": { "type": "number" }
      }
    }
  }
}